[713] [Time Beats: 52.04%] [Memory Beats: 65.26%] - LeetPush

+++ leetcode problem/713-Subarray-Product-Less-Than-K.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int numSubarrayProductLessThanK(vector<int>& nums, int k) {
         int left = 0 , right = 0 , cnt = 0 ; 
-        long long mul = 1;
+        int mul = 1;
         if(k <= 1){
             return 0;
         }
